feat: add photos-list and albums-list
